html_to_text: Extract code for html to plain text conversion.
